+++ src/Javis.jl
@@ -1,5 +1,6 @@
 module Javis
 
+using Animations
 using Cairo: CairoImageSurface, image
 using ColorTypes: ARGB32
 using FFMPEG
@@ -15,6 +16,18 @@ using Random
 using VideoIO
 
 const FRAMES_SYMBOL = [:same]
+
+struct ReversedEasing
+    easing::Easing
+end
+
+"""
+    rev(e::Easing)
+
+Reverse an easing function such that `easing_to_animation` maps it to `[1.0, 0.0]` instead of `[0.0, 1.0]`.
+An example can be seen in [`rotate`](@ref)
+"""
+rev(e::Easing) = ReversedEasing(e)
 
 """
     Video
@@ -139,6 +152,7 @@ A SubAction should not be created by hand but instead by using one of the constr
 
 # Fields
 - `frames::Frames`: the frames relative to the parent [`Action`](@ref)
+- `anim::Animation`: defines the interpolation function for the transitions
 - `func::Function`: the function that gets called in each of those frames.
     Takes the following arguments: `video, action, subaction, rel_frame`
 - `transitions::Vector{Transition}`: A list of transitions like [`Translation`](@ref)
@@ -147,13 +161,52 @@ A SubAction should not be created by hand but instead by using one of the constr
 """
 mutable struct SubAction <: AbstractAction
     frames::Frames
+    anim::Animation
     func::Function
     transitions::Vector{Transition}
     internal_transitions::Vector{InternalTransition}
 end
 
 """
-    SubAction(frames::UnitRange, func::Function)
+    SubAction(frames, easing::Union{ReversedEasing, Easing}, args...)
+
+A `SubAction` can be defined with frames an
+[easing function](https://jkrumbiegel.github.io/Animations.jl/stable/#Easings-1) and either
+a function or transformation(s).
+
+
+# Example
+In the following example a filled circle with radius 50 appears in the first 20 frames,
+which means the opacity is increased from 0 to 1.0. The interpolation function used here is
+sineio from [`Animations.jl`](https://github.com/jkrumbiegel/Animations.jl).
+Then it stays at full opacity and disappears the same way in the last 20 frames using a
+linear decay.
+
+```julia
+javis(demo, [
+    BackgroundAction(1:100, ground),
+    Action((args...)->circle(O, 50, :fill); subactions = [
+        SubAction(1:20, sineio(), appear(:fade)),
+        SubAction(81:100, disappear(:fade))
+    ])
+])
+```
+
+# Arguments
+- `frames`: A list of frames for which the function should be called.
+    - The frame numbers are relative to the parent [`Action`](@ref).
+- `easing::Union{ReversedEasing, Easing}`: The easing function for `args...`
+- `args...`: Either a function like [`appear`](@ref) or a Transformation
+    like [`Translation`](@ref)
+"""
+SubAction(frames, easing::Union{ReversedEasing,Easing}, args...) =
+    SubAction(frames, easing_to_animation(easing), args...)
+
+SubAction(frames, anim::Animation, transition::Transition...) =
+    SubAction(frames, anim, (args...) -> 1, transition...)
+
+"""
+    SubAction(frames, func::Function)
 
 A `SubAction` can be defined with frames and a function
 inside the `subactions` kwarg of an [`Action`](@ref).
@@ -178,7 +231,7 @@ javis(demo, [
     - For [`appear`](@ref) and [`disappear`](@ref) a closure exists,
       such that `appear(:fade)` works.
 """
-SubAction(frames, func::Function) = SubAction(frames, func, [], [])
+SubAction(frames, func::Function) = SubAction(frames, easing_to_animation(linear()), func)
 
 """
     SubAction(frames, trans::Transition...)
@@ -211,7 +264,11 @@ javis(demo, [
 - `trans::Transition...`: A list of transitions that shall be performed.
 """
 SubAction(frames, trans::Transition...) =
-    SubAction(frames, (args...) -> 1, collect(trans), [])
+    SubAction(frames, easing_to_animation(linear()), (args...) -> 1, trans...)
+
+
+SubAction(frames, anim::Animation, func::Function, transitions::Transition...) =
+    SubAction(frames, anim, func::Function, collect(transitions), [])
 
 """
     ActionSetting
@@ -274,6 +331,7 @@ Defines what is drawn in a defined frame range.
 - `id::Union{Nothing, Symbol}`: An id which can be used to save the result of `func`
 - `func::Function`: The drawing function which draws something on the canvas.
     It gets called with the arguments `video, action, frame`
+- `anim::Animation`: defines the interpolation function for the transitions
 - `transitions::Vector{Transition}` a list of transitions
     which can be performed before the function gets called.
 - `internal_transitions::Vector{InternalTransition}`:
@@ -286,6 +344,7 @@ mutable struct Action <: AbstractAction
     frames::Frames
     id::Union{Nothing,Symbol}
     func::Function
+    anim::Animation
     transitions::Vector{Transition}
     internal_transitions::Vector{InternalTransition}
     subactions::Vector{SubAction}
@@ -344,6 +403,9 @@ The current action can be accessed using CURRENT_ACTION[1]
 """
 const CURRENT_ACTION = Array{Action,1}()
 
+easing_to_animation(easing) = Animation(0.0, 0.0, easing, 1.0, 1.0)
+easing_to_animation(rev_easing::ReversedEasing) =
+    Animation(0.0, 1.0, rev_easing.easing, 1.0, 0.0)
 
 """
     Action(frames, func::Function, args...)
@@ -383,6 +445,52 @@ Action(func::Function, args...; kwargs...) =
     Action(frames, id::Union{Nothing,Symbol}, func::Function,
            transitions::Transition...; kwargs...)
 
+Fallback constructor for an Action which doesn't define an animation.
+A linear animation is assumed.
+"""
+function Action(
+    frames,
+    id::Union{Nothing,Symbol},
+    func::Function,
+    transitions::Transition...;
+    kwargs...,
+)
+
+    Action(frames, id, func, easing_to_animation(linear()), transitions...; kwargs...)
+end
+
+"""
+    Action(frames, id::Union{Nothing,Symbol}, func::Function, easing::Union{ReversedEasing, Easing},
+           args...; kwargs...)
+
+Fallback constructor for an Action which does define an animation using an easing function.
+
+# Example
+```
+javis(
+    video, [
+        BackgroundAction(1:100, ground),
+        Action((args...)->t(), sineio(), Translation(250, 0))
+    ]
+)
+```
+"""
+function Action(
+    frames,
+    id::Union{Nothing,Symbol},
+    func::Function,
+    easing::Union{ReversedEasing,Easing},
+    args...;
+    kwargs...,
+)
+
+    Action(frames, id, func, easing_to_animation(easing), args...; kwargs...)
+end
+
+"""
+    Action(frames, id::Union{Nothing,Symbol}, func::Function,
+           transitions::Transition...; kwargs...)
+
 # Arguments
 - `frames`: defines for which frames this action is called
 - `id::Symbol`: Is used if the `func` returns something which
@@ -397,6 +505,7 @@ function Action(
     frames,
     id::Union{Nothing,Symbol},
     func::Function,
+    anim::Animation,
     transitions::Transition...;
     kwargs...,
 )
@@ -407,7 +516,17 @@ function Action(
         subactions = opts[:subactions]
         delete!(opts, :subactions)
     end
-    Action(frames, id, func, collect(transitions), [], subactions, ActionSetting(), opts)
+    Action(
+        frames,
+        id,
+        func,
+        anim,
+        collect(transitions),
+        [],
+        subactions,
+        ActionSetting(),
+        opts,
+    )
 end
 
 """
@@ -719,9 +838,7 @@ function compute_transition!(
     action::AbstractAction,
     frame,
 )
-    t = (frame - first(get_frames(action))) / (length(get_frames(action)) - 1)
-    # makes sense to only allow 0 ≤ t ≤ 1
-    t = min(1.0, t)
+    t = get_interpolation(action, frame)
     from, to, center = rotation.from, rotation.to, rotation.center
 
     center isa Symbol && (center = pos(center))
@@ -748,9 +865,7 @@ function compute_transition!(
     action::AbstractAction,
     frame,
 )
-    t = (frame - first(get_frames(action))) / (length(get_frames(action)) - 1)
-    # makes sense to only allow 0 ≤ t ≤ 1
-    t = min(1.0, t)
+    t = get_interpolation(action, frame)
     from, to = translation.from, translation.to
 
     from isa Symbol && (from = pos(from))
@@ -1266,7 +1381,9 @@ export Line, Translation, Rotation, Transformation, Scaling
 export val, pos, ang, get_value, get_position, get_angle
 export get_javis_frame
 export projection, morph
-export appear, disappear
+export appear, disappear, rotate_around
+export rev
+export scaleto
 
 # custom override of luxor extensions
 export setline, setopacity, fontsize, get_fontsize, scale